+++ src/test/java/biz/gabrys/maven/plugin/util/timer/SystemTimerTest.java
@@ -14,24 +14,24 @@ public final class SystemTimerTest {
 
         timer.start();
         sleep();
-        final Time time1 = timer.getTime();
-        assertThat(time1.toMilliseconds()).isGreaterThan(0L);
+        final TimeSpan timeSpan1 = timer.getTime();
+        assertThat(timeSpan1.toMilliseconds()).isGreaterThan(0L);
 
         sleep();
-        final Time time2 = timer.getTime();
-        assertThat(time2.toMilliseconds()).isGreaterThan(0L);
-        assertThat(time1.toMilliseconds()).isLessThan(time2.toMilliseconds());
+        final TimeSpan timeSpan2 = timer.getTime();
+        assertThat(timeSpan2.toMilliseconds()).isGreaterThan(0L);
+        assertThat(timeSpan1.toMilliseconds()).isLessThan(timeSpan2.toMilliseconds());
 
         sleep();
-        final Time time3 = timer.stop();
-        assertThat(time2.toMilliseconds()).isLessThan(time3.toMilliseconds());
+        final TimeSpan timeSpan3 = timer.stop();
+        assertThat(timeSpan2.toMilliseconds()).isLessThan(timeSpan3.toMilliseconds());
 
         // getTime always return the same value after stop
-        assertThat(timer.getTime()).isSameAs(time3);
+        assertThat(timer.getTime()).isSameAs(timeSpan3);
         assertThat(timer.getTime()).isSameAs(timer.getTime());
 
         sleep();
-        assertThat(time3).isSameAs(timer.stop());
+        assertThat(timeSpan3).isSameAs(timer.stop());
     }
 
     private void sleep() throws InterruptedException {